Moved apStarWavegrid function to tools.__init__.py, and put appropriate import statement in spec.plot.py
